refactor : 화살표 버튼 이름 변경
